[#1381] neofs-cli: Move `control` command to a separate module
